pyproject.toml
[tool.poetry]
name = "schuaro"
version = "0.0.1"
description = ""
authors = ["Riley Wilton <meetingprince34@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.1"
redis = "^3.5.3"
Hypercorn = {extras = ["h3"], version = "^0.11.2"}
pymongo = "^3.11.4"
PyYAML = "^5.4.1"
trio = "^0.18.0"
databases = "^0.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
